Allow pod-to-pod communication across node boundaries in local dev setup.
The local dev setup using KinD has some implications with regards to the network setup.
In the outside layer, i.e. the kubernetes cluster acting as garden/seed, we use calico
with ip-in-ip overlay network. Previously, we used calico also for the inside layer, i.e.
the shoot kubernetes cluster, running in nodes, which are pods in the outside layer.
In principle, the pod-to-pod communication across node boundaries in the shoot cluster
should be easy to implement by either using an overlay network or plain ip routing in
the outside layer, i.e. without overlay network. Unfortunately, calico (in the outside
layer) prevents both approaches as it
1. drops overlay traffic coming from workload pods per default (can be disabled using
   `FelixConfiguration.spec.allowIPIPPacketsFromWorkloads`)
2. removes additional routes on calico devices (can be disabled using
   `FelixConfiguration.spec.removeExternalRoutes`)
3. prevents workloads from using IPs that differ from their assigned IP as spoofing
   protection
The first point makes it difficult to use a calico overlay network in the shoot. It is
possible, though, but requires to patch the `FelixConfiguration` when calico is running.
The second point makes running a non-overlay case hard. However, point three renders the
non-overlay scenario completely unusable in this scenario as no mitigation is available.

The rather simple solution to the double calico mess used in this change is to switch
the inside layer to use cilium instead of calico. With this approach, the shoot cluster
can use an overlay network in the inside layer, which is not detected/dropped by the
outside layer. Cilium uses vxlan on a different port than calico expects. Alternatively,
it is also possible to run cilium with geneve as tunnel protocol, which is not checked
by calico at all.

This approach makes it possible that shoot pods running on different virtual nodes can
flawlessly communicate with each other as you would expect. It works regardless whether
the outside layer has a single container/node or multiple as in the ha-multi-zone setup.

The end-to-end tests were also adapted to use cilium instead of calico.

It is still possible to use calico for shoot clusters in this setup, but the restriction
with regards to multiple nodes still applies there.