Switch back to kind node's hostname
We tried to use `kubernetes.default.svc` instead of the kind node's hostname
as the garden cluster address for the gardenlet values.
This works from within the kind cluster itself also if it is IPv6 single-stack.
However, it doesn't work from within `ManagedSeeds`.
Hence, the managed seed e2e fails because gardenlet cannot register itself in the garden cluster.