Adapt `NetworkPolicy`s for components in `garden` namespace (#7636)
* Enable `NetworkPolicy` controllers for `garden` namespace

* Add `NetworkPolicy`s for local setup to allow communication for etcd/control plane

* Add missing network policy labels for communication with runtime API server

Since `gardenlet` is the root component, it needs permissions for all networks. It will create a network policy specific for the seed's API server later, but we need the `0.0.0.0/0` rule for bootstrapping.

* Drop `allow-to-aggregate-prometheus` network policy

It was only used by Grafana, and it can use the new network policies instead.

* Drop `allow-to-seed-prometheus` network policy

It was only used by Grafana, and it can use the new network policies instead.

* Drop `allow-to-all-shoot-apiservers` network policy

It was only used by dependency-watchdog, and it can use the new network policies instead.

* Delete no longer required `global-network-policies` `ManagedResource` in `garden` namespace

* Drop `allow-to-loki` network policy

It was only used by Grafana, and it can use the new network policies instead.

* Replace custom rule for `fluent-bit` -> `loki`s traffic

fluent-bit talks to all lokis in the garden or shoot namespaces.

* Allow `aggregate-prometheus` to scrape its targets

* Allow `seed-prometheus` to scrape its targets

* Allow cache `prometheus` to scrape its targets

* Allow shoot `prometheus` to talk to cache `prometheus` and `alertmanager`

* Drop `allow-to-reversed-vpn-auth-server` network policy

* Extend GRM's `NetworkPolicy` controller to optionally watch `Ingress`es

* Provide ingress controller peer to GRM's `NetworkPolicy` controller

* Adapt `NetworkPolicy`s related to ingress-controller traffic

* Harmonize documentation

* Allow ports of webhooks in runtime cluster to be accessed from world

* GRM's `NetworkPolicy` controller no longer uses finalizers

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback