Remove provider-local limitation with regards to multiple nodes.
It is now possible to run shoot clusters with multiple nodes also with calico as CNI.
The felix configuration of the garden/seed cluster's calico configuration is adapted
to allow ip-in-ip packets originating from the workload. This allows use of calico
or cilium as shoot CNI. Both of them work with multiple nodes now. The default is
left at cilium for now.