Use `etcd-main` endpoint for full snapshots
Gardener used a PodExecutor to initiate full etcd snapshots. For performance and simplification reasons, it should rather call the HTTP endpoint of the `etcd-main` service directly. In addition, we drop dependencies on binaries that must be available in the `etcd-backup-restore` container, e.g. it was expected that `curl` was available in the container image.