Initialize connection to virtual cluster after KAPI is ready
The client will be needed for secrets rotation (and maybe more in the future).
For now, we'll misuse the static token kubeconfig until we deploy GRM with an enabled token-requestor controller that can generate a kubeconfig with a `12h`-valid token for GOP (similar to what happens for shoots for the gardenlet when it accesses shoots).