Fix the network policy for "cache Prometheus -> node-exporter" (#7772)
* Fix the network policy for "cache Prometheus -> node-exporter"

The node-exporter pod has "hostNetwork: true". The node-exporter is using the
host's network and it does not use the pod networking feature of Kubernetes.

Network policies do not seem to work in that case as intuitively expected, at
least with the Calico CNI plugin: even though the pod selector is correctly
configured in the network policy's egress/to section, the network path is still
blocked.

When the traffic is not restricted by the destination (only by the port), the
network path is allowed. Unfortunately it seems that (at least currently) the
destination can not be restricted by using pod selectors.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Address PR review feedback

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

---------

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>