Ensure consistency for the kube-apiserver service's networking annotation
Access to the kube-apiserver service in the control plane is governed by network
policies that are generated based on the annotations of the kube-apiserver
service.

See https://github.com/gardener/gardener/blob/31a167b526a3b9496830bcbb037cf470215291f7/docs/usage/network_policies.md#seed-cluster

When the gardenlet starts, it executes (1):

https://github.com/gardener/gardener/blob/375af561804331f88c6da9fa570a39a29bd48edb/cmd/gardenlet/app/app.go#L567

and when the shoot is reconciled, and if the `FullNetworkPoliciesInRuntimeCluster`
feature flag is enabled, the following lines are executed (2):

https://github.com/gardener/gardener/blob/1e26fbb41f0b84008d25665d421b4ab65cea3393/pkg/operation/botanist/component/kubeapiserverexposure/kube_apiserver_service.go#L188

In the local setup, where the `FullNetworkPoliciesInRuntimeCluster` feature flag
is enabled, I noticed that the kube-apiserver loses the
`{"matchLabels":{"gardener.cloud/role":"extension"}}` annotation which leads to
blocking the network path from the extensions to the kube-apiserver service when
the gardenlet is restarted. This eventually manifests in a failing
`EveryNodeReady` condition of the shoot. The annotation and the connectivity is
restored when the shoot is reconciled.

The reason for that is that the annotation that is set on startup in (1) and
upon reconciliation in (2) is not the same if the feature flag
`FullNetworkPoliciesInRuntimeCluster` is enabled: the `extension` selector is
missing in (1) in that case.

This PR adds the `extension` selector to (1), but only for the kube-apiserver
service which can be checked by the existing parameter
`withGardenNamespaceSelector`. It is not added for the
`vpn-seed-server services`.

The order of the selectors is adjusted for consistency with (2): the `garden`
namespace selector should come first.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>