Fix a minor formatting issue in the network policies annotation (#7788)
* Fix a minor formatting issue in the network policies annotation

The generated network policies have a `gardener.cloud/description` annotation
that contains a human readable description of the network policy.

For the `ingress-*-from-world` policies, a `kubectl get netpol -w` list+watch
request revealed that they are continuously updated (as observed in the local
setup). The slice of ports is formatted with `"%v"` for the description
annotation, but `fmt.Printf` did not use the generated `String()` function as
expected.

`fmt.Printf` used the default serialisation of the port struct, like
`{0x0123456789 10250 <nil>}`, and as the pointer to the protocol field changes
in each iteration, each iteration yielded a new string representation and hence
an update for the network policy, without no meaningful change.

The intention was to use the generated `String()` function for the
formatting, but it is not called for the `"%v"` printf verb for a slice of
values, because the `String()` function is defined for a pointer type
receiver.

```
  func (this *NetworkPolicyPort) String() string
```

See https://go.dev/doc/effective_go#printing

Instead of defining the (generated) `String()` function also for a value type
receiver, a utility function is added which copies the slice of values into a
new slice of pointers.

The new representation looks like this and does not contain randomly changing
addresses.

```
  [&NetworkPolicyPort{Protocol:*TCP,Port:10250,EndPort:nil,}]
```

Note that in general the NetworkPolicyPort can represent a port range, but it is
not really relevant in this use case, so in the next commit, only the ports are
going to be shown.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Use only the port number (or name) in the description

Instead of

```
  [&NetworkPolicyPort{Protocol:*TCP,Port:10250,EndPort:nil,}]
```

the description now contains only the port numbers (or names), like:

```
  [10250]
```

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Address PR review feedback

Show the Protocol as well.

Instead of `[10250]`, it now shows `[TCP/10250]` and handles the case
when the protocol is nil.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

---------

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>