Drop chart for provider-local `ControllerRegistration` and deploy it directly using skaffold (#7757)
* Don't set `version` and `appVersion` of helm chart

This is useless if not published to a proper helm repository

* `generate-controller-registration.sh` support fully-qualified image value

If the chart has a single fully-qualified `image` value instead of the usual
`image` stanza with explicit `repository` and `tag` fields, the script also
outputs a fully-qualified `image` value in the `ControllerDeployment`.

* Use fully-qualified `image` value in provider-local chart

* Drop chart for `ControllerRegistration` and deploy it directly using skaffold

This chart was used as a workaround for injecting the skaffold-built image into
the provider-local `ControllerDeployment`.
Now that the `ControllerDeployment` has a string `image` field instead of the
`image` stanza, we can instruct skaffold to overwrite that, just like it does
for other built-in API types like `Deployment`.

* Further simplify generation of provider-local `ControllerRegistration`

* Simplify setting of skaffold HA profiles

* Use same tools for both `generate*` targets

* Fix skaffold `provider-local` config for IPv6