Add support for configuring user-defined kernel settings per worker group. (#7825)
* Add support for configuring user-defined kernel settings per worker group.

Gardener tries hard to provide meaningful default values for kernel settings.
However, due to the vast breadth of use cases there will always be scenarios,
for which the default values might not be optimal. Overriding kernel settings
that are already defined by Gardener is not quite straight forward as Gardener
may re-apply its defaults at times. Therefore, it might be beneficial to
allow shoot owners to specify a set of kernel settings with specific values
to be applied by Gardener on each node of a worker group. This allows to have
custom kernel settings per worker group so that each worker group can be
optimized to its use case.

* Add sysctls to shoot example

* Addressed review feedback