[operator] Manage `istio` deployments (#7817)
* Change name suffix to prefix

This change will later be needed because Garden resources are typically
prefixed with `virtual-garden`.

* Fix Istio ManagedResource handling

Earlier, the Istio component did not consider all involved ManagedResources
when Wait, WainCleanup and Destroy were called.

* Create shared Istio component

- Move common Istio instantiation logic to `/pkg/operation/botanist/component/shared`
- Parameterize configuration differences between Seed and Garden cluster
- Adapt Seed controller to use shared Istio object factory

* Patch Status of Istio-Ingress Service

The Istio-Ingress `LoadBalancer` service must be patched in local development environments, similar to the Kube-Apiserver service.

* Consider resources in all namespaces

Gardener-Operator only acted in the `garden` namespace which is not true
anymore when Istio is used. Hence, the `garden` namespace option needs
to be removed from the Manager as well as cluster permissions need to be
widened.

* Move Istio helpers to shared package

Move Istio helpers to shared package to avoid multiple transitive package dependencies for Gardener-Operator that otherwise would need to be added to `skaffold-operator.yaml`.
Moving the functions to `pkg/utils/gardener` (as an alternative) is not possible because it results in cyclic dependency errors.

* Make GOP to deploy Istio

* Improve Seed health check

Consider 'SeedIsGarden' case also for health checks.

* Adjust docs

* Address review comments

* Address review comments II

* Address review comment III