Terminate gardenlet when seed gets registered as garden during runtime
If this happens, gardener-operator will run the NetworkPolicy controller and gardenlet should no longer run it. However, there is no way to stop a controller during runtime after it has been started. Hence, for now, our only option is to cancel the context of the manager which effectively stops gardenlet and restarts it.
In the "worst case", gardenlet continues to run the NetworkPolicy controller for another minute in parallel to the gardener-operator, however this should not be harmful and cannot be avoided anyways.