[operator] Enable `NetworkPolicy` protection in garden cluster (#7859)
* Introduce `FullNetworkPoliciesInRuntimeCluster` feature gate in `gardener-operator`

* Always include garden namespace and define ingress selector

This effectively enables the `NetworkPolicy` controller for the `garden` namespace also in garden clusters.

* Make gardenlet's `NetworkPolicy` controller reusable

* Expose runtime networking config in `Garden` resource

This will be needed as configuration for the `NetworkPolicy` controller (see latter commits)

* Add `NetworkPolicy` controller to `gardener-operator`

* Allow traffic for gardener-operator

Similar to
- https://github.com/gardener/gardener/pull/7636/commits/55319e154cb403430f639af4e1c20322c7cf314d
- https://github.com/gardener/gardener/pull/7714/commits/152330a95ab544eeaa4c0ffb1bd617b765485989#diff-3f3ce4c8d7c01a90835a7785fcaea4d2eb316a5b8e42cf709f4d56ccb6b63221R67

* Use `kubernetes.io/metadata.name` label for namespace selectors

In garden clusters, neither the `garden` nor the `kube-system` namespaces are labeled with the `role` key, so let's better always use the `kubernetes.io/metadata.name` key. It's more general anyways.

* Consider name prefix for `NetworkPolicy` label

`virtual-garden-kube-apiserver` talks to `virtual-garden-etcd-*`

* Remove workarounds for network policies

* Terminate gardenlet when seed gets registered as garden during runtime

If this happens, gardener-operator will run the NetworkPolicy controller and gardenlet should no longer run it. However, there is no way to stop a controller during runtime after it has been started. Hence, for now, our only option is to cancel the context of the manager which effectively stops gardenlet and restarts it.
In the "worst case", gardenlet continues to run the NetworkPolicy controller for another minute in parallel to the gardener-operator, however this should not be harmful and cannot be avoided anyways.

* Address PR review feedback

* Address PR review feedback