Stop the promtail service on the shoot nodes
The logs from the pods in the shoot cluster's kube-system namespace are sent to
Loki/Vali in the control plane (in the seed) by the promtail/valitail systemd
service that runs on each shoot node.

The promtail.service starts promtail, the auxiliary promtail-fetch-token.service
periodically fetches a bearer token that is used to authenticate promtail with
Loki.

This PR so far consistently renamed promtail to valitail, and hence when
applied, a new valitail.service and valitail-fetch-token.service will be added
to systemd on each shoot node. However, this generic rename approach does not
stop/clean up the old promtail systemd services which would continue to run on
the "updated" shoot nodes. The old promtail services would not "work" because
promtail is not compatible with Vali, but they would still run and unnecessarily
consume some memory (~50MB) on the shoot nodes.

This commit takes care of stopping the promtail services to avoid unnecessarily
using memory on the shoot nodes. Note that the promtail services are not
explicitly "removed" by this commit, they are just stopped: it is sufficient to
do so to prevent using resources unnecessarily. When the node is recreated due
to an OS/Kubernetes version upgrade or due to cluster autoscaling events, only
the new valitail services are going to be provisioned. In that pristine state,
upon startup, a single log line in journalctl will show that the valitail
service attempted to stop the no longer existing promtail services.

This commit can be reverted in a future release. Note that we need to stop the
promtail services only once on updated nodes, so it is fine to revert this
commit in the next Gardener release even if the shoot nodes are not recreated by
then.

Co-authored-by: Niki Dokovski <nickytd@gmail.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>