Do not add the DNS entry for shoot if it's `update-zone`
Do not add the entry for the e2e-upd-zone test as the target ip is dynamic.
The shoot cluster in e2e-upd-zone is created as single-zone control plane and afterwards updated to a multi-zone control plane.
This means that the external loadbalancer IP will change from a zone-specific istio ingress gateway to the default istio ingress gateway.
A static mapping (to the default istio ingress gateway) as done here will not work in this scenario.
The e2e-upd-zone test uses the in-cluster coredns for name resolution and can therefore resolve the api endpoint.