Push back the migration code removal to 6 releases in the future
That gives about 3 months time so that all the shoots on all the Gardener
landscapes can be migrated from Loki to Vali.

Note that the migration would happen in a healthy shoot in the very first
reconciliation window. If the shoot is broken (e.g. invalid shoot infrastructure
credentials) then the reconciliation will break on each attempt and that can
delay the migration from Loki to Vali indefinitely. We assume that 3 months are
sufficient so that eventually all the shoots are migrated and we can safely
remove the migration logic by that time.

This commit can be reverted in a future release.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>