Label shoot access secrets with `shoot` class
The controller in `gardener-resource-manager` is still not configured with the class, hence it is still responsible for all secrets.
This is fine for now since it is restricted to only the namespace it runs in, but in the future we should make sure that we restrict it to the `shoot` class to prevent potential future responsibility overlaps.