[provider-local] Drop `Migrate` overwrite in `Worker` controller
Earlier, the `Worker` controller of `provider-local` has overwritten the `Migrate` function which caused the essential migration logic in the generic `Worker` actuator to not get executed. This limited the accuracy of the CPM e2e test since the essential logic was not executed.
Similarly, `Restore` was not executing the essential logic of the generic `Worker` actuator. Now it is calling the the `RestoreWithoutReconcile()` and `Reconcile()` afterwards.

In the next commit, it will perform some custom logic to allow making use of the essential CPM logic in generic `Worker` actuator.

Co-Authored-By: Jens Schneider <schneider@23technologies.cloud>
Co-Authored-By: mreiger <michael@rauschpfeife.net>