Improve accuracy of local control plane migration e2e tests (#7981)
* Decouple certain functions from generic `Worker` actuator

This way, they become reusable even when no `genericActuator` object is available

Co-Authored-By: Jens Schneider <schneider@23technologies.cloud>
Co-Authored-By: mreiger <michael@rauschpfeife.net>

* Move restoration logic in generic `Worker` actuator into dedicated function

Earlier, the `genericactuator.Restore()` was restoring and always calling `Reconcile()` at the end. It is still doing this, however now the restoration logic has been separated into `RestoreWithoutReconcile`. This allows extensions to optionally perform custom operations before they call the `Reconcile()` function.

Co-Authored-By: Jens Schneider <schneider@23technologies.cloud>
Co-Authored-By: mreiger <michael@rauschpfeife.net>

* [provider-local] Drop `Migrate` overwrite in `Worker` controller

Earlier, the `Worker` controller of `provider-local` has overwritten the `Migrate` function which caused the essential migration logic in the generic `Worker` actuator to not get executed. This limited the accuracy of the CPM e2e test since the essential logic was not executed.
Similarly, `Restore` was not executing the essential logic of the generic `Worker` actuator. Now it is calling the the `RestoreWithoutReconcile()` and `Reconcile()` afterwards.

In the next commit, it will perform some custom logic to allow making use of the essential CPM logic in generic `Worker` actuator.

Co-Authored-By: Jens Schneider <schneider@23technologies.cloud>
Co-Authored-By: mreiger <michael@rauschpfeife.net>

* [provider-local] Delete stale `Machine`s after CPM

Co-Authored-By: Jens Schneider <schneider@23technologies.cloud>
Co-Authored-By: mreiger <michael@rauschpfeife.net>

* Address PR review feedback

---------

Co-authored-by: Jens Schneider <schneider@23technologies.cloud>
Co-authored-by: mreiger <michael@rauschpfeife.net>