Increased default reconciliation timeout to 3 minutes.
The default reconciliation timeout is used as a safety net to prevent reconciliation
operations to hang forever. Therefore, it should be long enough that all usual
operations succeed, but short enough so that the waiting time is not endless in case
of issues/hanging operations.
With larger amount of resources being applied for example as managed resources the
default reconciliation timeout needs to be increased. An example for such a resource
is the istio ingress gateway in case the seed uses multiple availability zones as
the managed resource will then include all entities for all availability zones.
Another contributing factor may be if the control plane of a seed is latency-wise
fairly distant to its data plane causing operations such as applying resources to
be rather slow.