Sync client CA to `InternalSecret` and use it in `shoots/adminkubeconfig` subresource (#8088)
* Sync `ca-client` to `InternalSecret` in project namespace

* Handle `InternalSecrets` in `Seed{Authorizer,Restriction}`

* Prefactor: pass informer factories to API server registry

Currently, the `shoots/adminkubeconfig` storage extracts the `ca-client` from the `ShootState` object directly from `etcd` (passes around a `*genericregistry.Store`).
This works but is undesired. The subresource should use informers/listers instead.
For now, we will keep the current mechanism for retrieving `ShootStates` (which will be dropped eventually), but we will use listers in the new approach for retrieving `InternalSecrets`.

For this, this commits fixes the initialization sequence for the API server's informers.
Currently, they are only used in admission plugins but they are initialized too late for passing them to the registry.
There's no logical change in this commit. We only change the order of initialization and pass on the informer factories.

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Adapt `shoots/adminkubeconfig` subresource to not fetch CAs via `ShootState`

With this commit, we fetch the `ca-cluster` and `ca-client` data from the project `Secret` and `InternalSecret` respectively.
For migration purposes, we fall back to the `ShootState`, if the `ca-client` has not been synced to the garden cluster.

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Add back `SeedAdmission` call for `DELETE backupbuckets`

* Address review suggestions

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>