Make `tokenrequestor` controller reusable and add it to gardenlet (#8032)
* Move `tokenrequestor` package from `pkg/resourcemanager/controller` to `pkg/controller`

It will be reused by `gardenlet` in the next commits, so let's move it to a neutral place.

* Introduce class for `tokenrequestor` controller

This class allows to run multiple instances of the controller and differentiate which of them is responsible by labelling the `Secret`s with `resources.gardener.cloud/class=<class>`.

* Label shoot access secrets with `shoot` class

The controller in `gardener-resource-manager` is still not configured with the class, hence it is still responsible for all secrets.
This is fine for now since it is restricted to only the namespace it runs in, but in the future we should make sure that we restrict it to the `shoot` class to prevent potential future responsibility overlaps.

* Make `TargetNamespace` configurable in `tokenrequestor` controller

In gardenlet, we don't want to allow requesting tokens for `ServiceAccounts` in arbitrary namespaces.
Instead, we want to use the seed's namespace in the garden cluster for all `ServiceAccounts`.

* Make `Audiences` configurable in `TokenRequestor` controller

While we know the audiences in shoot clusters (since `gardenlet` deploys the API server), we don't know them for the garden cluster.
Hence, we cannot set them.

* Extend `gardenlet`'s component config for `TokenRequestor` controller

The `gardenlet` will also add the `TokenRequestor` controller to its manager in order to request tokens for components running in the seed cluster which need to communicate with the garden cluster.

* Register `TokenRequestor` controller in gardenlet

* gardenlet may perform all operations for `ServiceAccount`s in seed namespace

* Prefactor: restructure project e2e tests

This package will be used as a template for the next commit. Restructure it a bit beforehand.

* Add e2e test for garden access secrets

* Address review suggestions

* Address review suggestions

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>