Fix flaky gardener operator tests caused by concurrent writes to garden object. (#8104)
Gardener operator tests sometimes failed when accessing the virtual kube-apiserver.
The result was usually a connection reset. This was due to the fact that istio
closed the corresponding port as it was reconfigured to use a different host name
for the gateway resource and hence had no matching virtual service/destination rule.
The original reason for the different host name with a missing suffix was probably
the concurrent update of the garden resource due to a parallel write (see
https://github.com/gardener/gardener/blob/master/docs/development/kubernetes-clients.md).
Therefore, the domain name is now calculated ahead of use during initialization
when the garden object is not changed concurrently.

Co-authored-by: Johannes Scheerer <johannes.scheerer@sap.com>