Allow deletion of shoot clusters with incorrect networking settings.
Previously, gardener allowed overlap of shoot network ranges and the vpn. This
was rather a bug than a feature, but led to shoot clusters in the wild, which
have such an overlap. Those clusters cannot be deleted at the moment as the
admission validation will reject them.
This change is to allow deletion of those clusters without weakening the general
validation for create/update.