[GEP-22] Drop `shootstate-extensions` and `shootstate-secret` controllers (#8136)
* Drop legacy `ShootState` controllers

* Remove legacy ShootState controller finalizer from secrets

* Drop {Get,Set}ShootState methods on `Botanist`

* Drop `EnsureShootStateExists` function

Instead, we now fetch the `ShootState` object in the restore phase only (this is the only phase when we need it) and store it directly on the internal `Shoot` object.

* Delete `ShootState` after restoration

Only if the gardenlet is responsible for an unmanaged `Seed` - otherwise, its `shoot-state` reconciler wants to perform regular backups of the `ShootState`, so we shouldn't delete it even after successful restoration.

* Disable `shoot-state` reconciler locally

From https://github.com/gardener/gardener/pull/8112#discussion_r1236417469:

> In our e2e tests the cpm test shoots are always scheduled on a non-managed seed. With this, the controller would always be active and we wouldn't verify that migration also works without periodic backups.

Since we don't need this periodic backup functionality in e2e tests, we simply disable the controller.