Prevent concurrent updates on in-memory Garden object of operator during parallel task flow. (#8145)
Patching/updating object via controller runtime kubernetes clients can have negative
consequences for parallel readers of the in-memory golang object (see
https://github.com/gardener/gardener/blob/master/docs/development/kubernetes-clients.md
for details). For example, parts of the structure can be temporarily (and unexpectedly)
empty.
In PR #8103, the immediate issue was addressed while the overarching problem remained.
This change reverts the hotfix and replaces it with a proper solution, i.e. it decouples
the continuous status patch updates of the progress reporter from the golang in-memory
object used during the parallel task flow. Therefore, reading the in-memory object should
now be safe again while the in-memory object is updated after the flow so that execution
can continue as usual.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>