Adapt `Worker` controller to persist state in `Migrate` phase (#8146)
* Move Worker state persist logic into function

Such that we can call it in Migrate() in the next commit

* Persist worker state on-demand

- In `Migrate()`, the generic Worker actuator persists the worker state before shallow-deleting the machine resources.
- In `Restore()`, the generic Worker actuator will set the persisted state to `nil` after successful restoration.

Since gardenlet is not yet persisting the `ShootState` only after all extension resources have been migrated, we have to keep the state reconciler for a bit longer. However, with this commit, revendored extensions are already prepared to persist their state in the `Migrate()` step as documented in the extensions contract in https://github.com/gardener/gardener/blob/master/docs/extensions/migration.md#migrate-and-restore-operations.

In the future, when `gardenlet` persists the `ShootState` only after all extension resources have been migrated, we can simply drop the state reconciler and clean the `.status.state` field again after successful restoration.