Renew all garden access secrets on seed when triggered by annotation (#8152)
* Nit: move the `tokenrequestor` controller docs out of the `shoot` controller section

* Add documentation for the new feature

* Validate operation annotation on `Seeds`

* Bump `Seed` generation when the operation annotation is set

* Allow selecting class in `RenewAccessSecrets`

With this, we can renew the garden access secrets in all namespaces in the next commit without renewing shoot access secrets.

In the future, the shoot/garden controller should also restrict the call to the `shoot` class.
However, the controller in `gardener-resource-manager` is still not configured with the class, hence it is still responsible for all secrets.
This is fine for now since resource-manager and the call to `RenewAccessSecrets` are restricted to the `garden` or shoot namespace, but in the future we should make sure that we restrict it to the `shoot` class to prevent potential future responsibility overlaps.

* Renew all garden access secrets on seed when triggered by annotation

* Verify garden access token renewal in e2e tests

* Address review suggestions