[operator] Introduce `dns.domains` field (#8156)
* Add new `dns.domains` field

Add to: Add new field

* Allow unsetting `dns.domain`

* Deprecate `dns.domain` field

* Use new `dns.domains` field

* Validate that domain is configured

* Always consider `gardener.` domain

The `gardener.` domain was previously always considered as an alternative domain name in the API server certificate. Consequently, the SNI exposure configuration for Istio should also include this seconds domain.

* Extract virtual cluster update validation

* Fix bug in update validation

* Enhance validation

Let's make the first configured domain in 'domains' immutable. Otherwise the service account issuer changes for the cluster which requires to re-issue all service account tokens. The 'domain' field is already immutable today, so we can still consider adding support for this use-case in the future.

* run [make generate]

* Address review comment

* Address review comment