Clean up DNS handling (#8199)
* Drop `IncludeZones` and `ExcludeZones`

`IncludeZones` and `ExcludeZones` are a leftover when Gardenlet created `DNSEntries` instead of `DNSRecords`.

* Drop `IncludeDomains` and `ExcludeDomains`

`IncludeDomains` and `ExcludeDomains` are obsolete and not supported by `DNSRecord`s.

* Drop irrelevant condition in shoot flow

Earlier, we used `!useSNI` which was later replaced by `!useDNS` (v1.73) but at those places a 1:1 replacement is not possible. We don't support non-SNI and thus won't deploy Kube-Apiserver Services of Type `LoadBalancer`. Now we can always deploy the service, also when the cluster is hibernated.

* Drop `nip.io` support

Support for `nip.io` is obsolete:
- It's been broken for a long time (shoot fails with `Could not initialize a new operation for Shoot cluster: unable to figure out which secret should be used for dns`)
- It's not documented at all

* [make generate]

* Address review comment

* Address review comments

* Deprecate 'Domains' and 'Zones'