Handle enablement of seed authorization/restriction features
The reconcile flow deploys the kube-apiserver of the virtual garden cluster before the gardener-apiserver and
gardener-admission-controller (it has to be this way, otherwise the Gardener components cannot start). However,
GAC serves an authorization webhook for the SeedAuthorizer feature. We can only configure kube-apiserver to
consult this webhook when GAC runs, obviously. This is not possible in the initial Garden deployment (due to
above order). Hence, we have to run the flow as second time after the initial Garden creation - this time with
the SeedAuthorizer feature getting enabled. From then on, all subsequent reconciliations can always enable it and
only one reconciliation is needed.