Fix CA rotation for multi-node `etcd`s (#8303)
Earlier, the peer client/server certificate was only signed by the current
CA if the rotation phase was in `Preparing`. This will become an issue
if a 'normal' reconciliation is triggered while the cluster is in `Prepared`
because then the certificate is singed by the old CA again. The next phase,
`Completing`, will eventually fail because `etcd` peers cannot validate their
certificates anymore.