Fix `gardensystem/virtual` component
It's managed resources should be created in the target cluster (virtual), not in the runtime cluster.

Also, keep `garden` namespace in virtual system `ManagedResource`: When deleting a `Garden` we shouldn't care about whether the `garden`
namespace in the virtual cluster still exists - let's just keep it and proceed with the deletion.