Allow the alertmanager in the garden ns to send out email notifications (#8316)
The alertmanager in the garden namespace of a seed can be configured to send
email notifications about the alerts of the shoot control plane Prometheus-es
and the alerts of the aggregate Prometheus in the garden namespace.

See `charts/gardener/controlplane/charts/runtime/values.yaml`,
`global.alerting`.

We noticed it with quite some delay that probably since the recent improvements
of the network policies, the `garden/alertmanager-0` pod does not have DNS and
public/private network access and hence could not send out the email
notifications.

```
k logs -n garden alertmanager-0 | tail -n 1

ts=2023-08-04T12:16:18.450Z
caller=dispatch.go:354
level=error
component=dispatcher
msg="Notify for alerts failed"
num_alerts=1
err="email-kubernetes-ops/email[0]:
     notify retry canceled after 2 attempts:
     establish connection to server:
     dial tcp 52.x.y.z:587: i/o timeout"
```

This change grants DNS and public/private network access to the alertmanager in
the garden namespace so that it can successfully send email notifications.

Note that the mail server could be either in a public or in a private network,
hence both network paths: to the public and to the private network need to be
allowed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>